% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piplines.R
\name{combine_countries}
\alias{combine_countries}
\title{Combines 2 couns in a vector}
\usage{
combine_countries(couns)
}
\arguments{
\item{couns}{The incoming country list}
}
\value{
\code{c(couns, couns)}
}
\description{
Combines 2 couns in a vector
}
