% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piplines.R
\name{get_pipeline}
\alias{get_pipeline}
\title{Creates a pipeline programmatically}
\usage{
get_pipeline(which_countries)
}
\arguments{
\item{which_countries}{The countries of interest for this pipeline}
}
\value{
A \code{targets} pipeline.
}
\description{
Creates a pipeline programmatically
}
